requirements.txt
fastapi
uvicorn[standard]
obsws-python
python-dotenv
python-multipart
google-genai
openai
requests
